+++ conf/appearance/themes/Dark+/script/module/config.js
@@ -270,6 +270,16 @@ export var config = {
                 enable: true, // 是否启用 iframe 重新加载
             },
         },
+        fullscreen: {
+            // 全屏
+            enable: true, // 是否启用全屏
+            iframe: {
+                enable: true, // 是否启用 iframe/widgets/video 全屏
+            },
+            mark: {
+                enable: true, // 是否开启双击块标全屏某个块
+            },
+        },
         doc: {
             enable: true, // 是否启用文档扩展功能
             heading: { // 标题操作
@@ -562,17 +572,30 @@ export var config = {
                         index: 5,
                     },
                     random: true, // 是否随机切换网络背景图片 URL
-                    light: [ // 随机亮色背景图片 URL
-                        'https://source.unsplash.com/random/1920x1080/?bright',
-                        'https://api.dujin.org/bing/1920.php',
-                        'https://unsplash.it/1920/1080?random',
-                        // 'https://api.ixiaowai.cn/gqapi/gqapi.php⁠⁠⁠⁠⁠⁠',
-                    ],
-                    dark: [ // 随机暗色背景图片 URL
-                        'https://source.unsplash.com/random/1920x1080/?night',
-                        'https://source.unsplash.com/random/1920x1080/?starry',
-                        'https://source.unsplash.com/random/1920x1080/?dark',
-                    ],
+                    landscape: { // 横屏背景图片
+                        light: [ // 随机亮色背景图片 URL
+                            'https://source.unsplash.com/random/1920x1080/?bright',
+                            'https://api.dujin.org/bing/1920.php',
+                            'https://unsplash.it/1920/1080?random',
+                            // 'https://api.ixiaowai.cn/gqapi/gqapi.php⁠⁠⁠⁠⁠⁠',
+                        ],
+                        dark: [ // 随机暗色背景图片 URL
+                            'https://source.unsplash.com/random/1920x1080/?night',
+                            'https://source.unsplash.com/random/1920x1080/?starry',
+                            'https://source.unsplash.com/random/1920x1080/?dark',
+                        ],
+                    },
+                    portrait: { // 竖屏背景图片
+                        light: [
+                            'https://source.unsplash.com/random/1080x1920/?bright',
+                            'https://unsplash.it/1080/1920?random',
+                        ],
+                        dark: [
+                            'https://source.unsplash.com/random/1080x1920/?night',
+                            'https://source.unsplash.com/random/1080x1920/?starry',
+                            'https://source.unsplash.com/random/1080x1920/?dark',
+                        ],
+                    },
                 },
                 custom: {
                     enable: true, // 自定义背景图片
@@ -590,14 +613,26 @@ export var config = {
                     },
                     random: true, // 是否随机选择自定义背景图片
                     default: false, // 是否默认使用自定义背景图片
-                    light: [ // 自定义亮色背景图片 URL 列表
-                        `${THEME_PATHNAME}/image/light/background-main.jpg`,
-                        `${THEME_PATHNAME}/image/light/background-dialog.jpg`,
-                    ],
-                    dark: [ // 自定义暗色背景图片 URL 列表
-                        `${THEME_PATHNAME}/image/dark/background-main.jpg`,
-                        `${THEME_PATHNAME}/image/dark/background-dialog.jpg`,
-                    ],
+                    landscape: { // 横屏背景图片
+                        light: [ // 自定义亮色背景图片 URL 列表
+                            `${THEME_PATHNAME}/image/light/background-main.jpg`,
+                            `${THEME_PATHNAME}/image/light/background-dialog.jpg`,
+                        ],
+                        dark: [ // 自定义暗色背景图片 URL 列表
+                            `${THEME_PATHNAME}/image/dark/background-main.jpg`,
+                            `${THEME_PATHNAME}/image/dark/background-dialog.jpg`,
+                        ],
+                    },
+                    portrait: { // 竖屏背景图片
+                        light: [ // 自定义亮色背景图片 URL 列表
+                            `${THEME_PATHNAME}/image/light/background-main-portrait.jpg`,
+                            `${THEME_PATHNAME}/image/light/background-dialog.jpg`,
+                        ],
+                        dark: [ // 自定义暗色背景图片 URL 列表
+                            `${THEME_PATHNAME}/image/dark/background-main-portrait.jpg`,
+                            `${THEME_PATHNAME}/image/dark/background-dialog.jpg`,
+                        ],
+                    },
                 },
             },
         },
@@ -1095,6 +1130,106 @@ export var config = {
                             },
                             {
                                 enable: true,
+                                type: { NodeDocument: { enable: true } },
+                                mode: "button",
+                                icon: "#iconRefresh",
+                                label: {
+                                    zh_CN: "重启内核",
+                                    other: "Restart Kernel",
+                                },
+                                click: {
+                                    enable: true,
+                                    callback: null,
+                                    tasks: [
+                                        {
+                                            type: 'jupyter-restart-kernel',
+                                            params: {},
+                                        },
+                                    ],
+                                },
+                            },
+                            {
+                                enable: true,
+                                type: { NodeDocument: { enable: true } },
+                                mode: "button",
+                                icon: "#iconPlay",
+                                label: {
+                                    zh_CN: "运行所有单元格 (转义输出: ✔ 控制字符: ✔)",
+                                    other: "Run All Cells (Escape: ✔ cntrl: ✔)",
+                                },
+                                click: {
+                                    enable: true,
+                                    callback: null,
+                                    tasks: [
+                                        {
+                                            type: 'jupyter-run-all-cells',
+                                            params: { escaped: true, cntrl: true },
+                                        },
+                                    ],
+                                },
+                            },
+                            {
+                                enable: true,
+                                type: { NodeDocument: { enable: true } },
+                                mode: "button",
+                                icon: "#iconPlay",
+                                label: {
+                                    zh_CN: "运行所有单元格 (转义输出: ✔ 控制字符: ✖)",
+                                    other: "Run All Cells (Escape: ✔ cntrl: ✖)",
+                                },
+                                click: {
+                                    enable: true,
+                                    callback: null,
+                                    tasks: [
+                                        {
+                                            type: 'jupyter-run-all-cells',
+                                            params: { escaped: true, cntrl: false },
+                                        },
+                                    ],
+                                },
+                            },
+                            {
+                                enable: true,
+                                type: { NodeDocument: { enable: true } },
+                                mode: "button",
+                                icon: "#iconPlay",
+                                label: {
+                                    zh_CN: "运行所有单元格 (转义输出: ✖ 控制字符: ✔)",
+                                    other: "Run All Cells (Escape: ✖ cntrl: ✔)",
+                                },
+                                click: {
+                                    enable: true,
+                                    callback: null,
+                                    tasks: [
+                                        {
+                                            type: 'jupyter-run-all-cells',
+                                            params: { escaped: false, cntrl: true },
+                                        },
+                                    ],
+                                },
+                            },
+                            {
+                                enable: true,
+                                type: { NodeDocument: { enable: true } },
+                                mode: "button",
+                                icon: "#iconPlay",
+                                label: {
+                                    zh_CN: "运行所有单元格 (转义输出: ✖ 控制字符: ✖)",
+                                    other: "Run All Cells (Escape: ✖ cntrl: ✖)",
+                                },
+                                click: {
+                                    enable: true,
+                                    callback: null,
+                                    tasks: [
+                                        {
+                                            type: 'jupyter-run-all-cells',
+                                            params: { escaped: false, cntrl: false },
+                                        },
+                                    ],
+                                },
+                            },
+                            {
+                                enable: true,
                                 type: { NodeCodeBlock: { enable: true, subtype: { null: true } } },
                                 mode: "button",
                                 icon: "#iconPlay",
@@ -1107,7 +1242,7 @@ export var config = {
                                     callback: null,
                                     tasks: [
                                         {
-                                            type: 'jupyter-run-code',
+                                            type: 'jupyter-run-cell',
                                             params: { escaped: true, cntrl: true },
                                         },
                                     ],
@@ -1127,7 +1262,7 @@ export var config = {
                                     callback: null,
                                     tasks: [
                                         {
-                                            type: 'jupyter-run-code',
+                                            type: 'jupyter-run-cell',
                                             params: { escaped: true, cntrl: false },
                                         },
                                     ],
@@ -1147,7 +1282,7 @@ export var config = {
                                     callback: null,
                                     tasks: [
                                         {
-                                            type: 'jupyter-run-code',
+                                            type: 'jupyter-run-cell',
                                             params: { escaped: false, cntrl: true },
                                         },
                                     ],
@@ -1167,7 +1302,7 @@ export var config = {
                                     callback: null,
                                     tasks: [
                                         {
-                                            type: 'jupyter-run-code',
+                                            type: 'jupyter-run-cell',
                                             params: { escaped: false, cntrl: false },
                                         },
                                     ],
@@ -1639,6 +1774,27 @@ export var config = {
                             },
                             {
                                 enable: true,
+                                type: null,
+                                mode: "button",
+                                icon: "#iconFullscreen",
+                                label: {
+                                    zh_CN: "全屏显示",
+                                    other: "Full-screen Display",
+                                },
+                                accelerator: () => config.theme.hotkeys.fullscreen.iframe,
+                                click: {
+                                    enable: true,
+                                    callback: null,
+                                    tasks: [
+                                        {
+                                            type: 'full-screen',
+                                            params: {},
+                                        },
+                                    ],
+                                },
+                            },
+                            {
+                                enable: true,
                                 type: {
                                     NodeDocument: { enable: true },
                                     NodeBlockQueryEmbed: { enable: true },
@@ -1702,7 +1858,7 @@ export var config = {
                                     NodeTable: { enable: true },
                                 },
                                 mode: "button",
-                                icon: "#iconInsertColumn",
+                                icon: "#iconEdit",
                                 label: {
                                     zh_CN: "切换书写模式",
                                     other: "Toggle Writing Modes",
@@ -2025,7 +2181,7 @@ export var config = {
                                 mode: "separator",
                             },
                             {
-                                enable: false,
+                                enable: false, // deprecated
                                 type: {
                                     NodeBlockQueryEmbed: { enable: true },
                                 },
@@ -2401,13 +2557,33 @@ export var config = {
                     key: 'F5',
                 },
                 iframe: {
-                    // 刷新 iframe 块(Ctrl + 单击)
+                    // 刷新 iframe/widgets 块 (Ctrl + 单击)
                     enable: true,
                     CtrlCmd: true,
                     WinCtrl: false,
                     Shift: false,
                     Alt: false,
                     type: 'click',
+                },
+            },
+            fullscreen: {
+                iframe: {
+                    // 全屏 iframe/widgets 块 (双击)
+                    enable: true,
+                    CtrlCmd: false,
+                    WinCtrl: false,
+                    Shift: false,
+                    Alt: false,
+                    type: 'dblclick',
+                },
+                mark: {
+                    // 全屏指定块 (双击块标)
+                    enable: true,
+                    CtrlCmd: false,
+                    WinCtrl: false,
+                    Shift: false,
+                    Alt: false,
+                    type: 'dblclick',
                 },
             },
             doc: {